+++ pytorch_helpers/habana_helpers/logging.cpp
@@ -156,7 +156,9 @@ std::string synStatusToStr(synStatus statusArg) {
 
 std::string formatStatusMsg(synStatus statusArg) {
   return fmt::format(
-      "synStatus {} [{}]. ", statusArg, synStatusToStr(statusArg));
+      "synStatus {} [{}]. ",
+      static_cast<int>(statusArg),
+      synStatusToStr(statusArg));
 }
 
 uint64_t get_tid_internal() {
